Refactor populateFetch in its own module
