+++ packages/strapi-connector-bookshelf/lib/utils/attributes.js
@@ -3,9 +3,9 @@
 /**
  * Returns the attribute keys of the component related attributes
  */
-function getComponentAttributes(attributes) {
-  return Object.keys(attributes).filter(key =>
-    ['component', 'dynamiczone'].includes(attributes[key].type)
+function getComponentAttributes(definition) {
+  return Object.keys(definition.attributes).filter(key =>
+    ['component', 'dynamiczone'].includes(definition.attributes[key].type)
   );
 }
